test(app): add local postgres test workflow
