+++ crates/coral-app/tests/grpc/server_lifecycle_tests.rs
@@ -13,7 +13,6 @@ use coral_client::{
     AppClient, default_workspace,
     local::{LocalServerError, ServerBuilder},
 };
-use sqlx::postgres::PgPoolOptions;
 use sqlx::sqlite::{SqliteConnectOptions, SqlitePoolOptions};
 use tempfile::TempDir;
 use tonic::Request;
@@ -88,31 +87,6 @@ async fn server_lifecycle_rejects_postgres_config_without_url_env_value() {
     }
 }
 
-#[tokio::test]
-#[ignore = "set CORAL_TEST_POSTGRES_URL to run configured Postgres startup coverage"]
-async fn server_lifecycle_can_start_with_postgres_database_config() {
-    let Some(database_url) = postgres_test_url() else {
-        return;
-    };
-    let temp = TempDir::new().expect("temp dir");
-    let config_dir = temp.path().join("coral-config");
-    fs::create_dir_all(&config_dir).expect("create config dir");
-    fs::write(
-        config_dir.join("config.toml"),
-        "[database]\nbackend = \"postgres\"\nurl_env = \"CORAL_TEST_POSTGRES_URL\"\n",
-    )
-    .expect("write config");
-
-    let server = ServerBuilder::new()
-        .with_config_dir(&config_dir)
-        .start()
-        .await
-        .expect("start server with Postgres config");
-    assert_postgres_db_is_migrated(&database_url).await;
-
-    server.shutdown().await.expect("shutdown server");
-}
-
 async fn assert_default_sqlite_db_is_migrated(config_dir: &std::path::Path) {
     let database_file = config_dir.join("coral.db");
     assert!(
@@ -132,27 +106,4 @@ async fn assert_default_sqlite_db_is_migrated(config_dir: &std::path::Path) {
     .await
     .expect("inspect migrated schema");
     assert_eq!(table_count, 1, "workspaces table should be migrated");
-}
-
-async fn assert_postgres_db_is_migrated(database_url: &str) {
-    let pool = PgPoolOptions::new()
-        .connect(database_url)
-        .await
-        .expect("open Postgres database");
-    let table_exists: bool =
-        sqlx::query_scalar("SELECT to_regclass('public.workspaces') IS NOT NULL")
-            .fetch_one(&pool)
-            .await
-            .expect("inspect migrated Postgres schema");
-    assert!(table_exists, "workspaces table should be migrated");
-}
-
-#[expect(
-    clippy::disallowed_methods,
-    reason = "The ignored Postgres integration test is explicitly gated by this CI/test-only variable."
-)]
-fn postgres_test_url() -> Option<String> {
-    std::env::var("CORAL_TEST_POSTGRES_URL")
-        .ok()
-        .filter(|value| !value.is_empty())
 }
